requirements.txt
aiohttp==3.7.4.post0
aiofiles==22.1.0
aiopg==1.4.0
anyio==3.5.0
fastapi==0.88.0
gunicorn==20.1.0
google-cloud-storage==2.13.0
pydantic==1.9.1
python-dotenv==0.21.0
python-multipart==0.0.5
psycopg2-binary==2.9.6
uvicorn[standard]==0.20.0
sqlalchemy==1.4.45